# apis/backend-api/content/news-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork News Backend API",
    "version": "1.0.0",
    "x-sdkwork-owner": "sdkwork-news",
    "x-sdkwork-api-authority": "sdkwork-news.backend"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080"
    }
  ],
...
  "paths": {
...
    "/backend/v3/api/news/items/{itemId}/publish": {
      "post": {
        "tags": [
          "news"
        ],
        "summary": "News items.publish",
        "operationId": "items.publish",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NewsItem"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-news",
        "x-sdkwork-api-authority": "sdkwork-news.backend",
        "x-sdkwork-domain": "news",
        "x-sdkwork-resource": "items",
        "x-sdkwork-public": false
      }
    },
...
    "/backend/v3/api/news/alerts/breaking/{alertId}/publish": {
      "post": {
        "tags": [
          "news"
        ],
        "summary": "News alerts.breaking.publish",
        "operationId": "alerts.breaking.publish",
        "parameters": [
          {
            "name": "alertId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NewsBreakingAlert"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-news",
        "x-sdkwork-api-authority": "sdkwork-news.backend",
        "x-sdkwork-domain": "news",
        "x-sdkwork-resource": "alerts",
        "x-sdkwork-public": false
      }
    },
...
    "/backend/v3/api/news/digests/{digestId}/publish": {
      "post": {
        "tags": [
          "news"
        ],
        "summary": "News digests.publish",
        "operationId": "digests.publish",
        "parameters": [
          {
            "name": "digestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NewsDigestIssue"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-news",
        "x-sdkwork-api-authority": "sdkwork-news.backend",
        "x-sdkwork-domain": "news",
        "x-sdkwork-resource": "digests",
        "x-sdkwork-public": false
      }
    },
...
    "/backend/v3/api/news/fact_checks/{factCheckId}/publish": {
      "post": {
        "tags": [
          "news"
        ],
        "summary": "News factChecks.publish",
        "operationId": "factChecks.publish",
        "parameters": [
          {
            "name": "factCheckId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NewsFactCheck"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-news",
        "x-sdkwork-api-authority": "sdkwork-news.backend",
        "x-sdkwork-domain": "news",
        "x-sdkwork-resource": "factChecks",
        "x-sdkwork-public": false
      }
    },
...
    "/backend/v3/api/news/corrections/{correctionId}/publish": {
      "post": {
        "tags": [
          "news"
        ],
        "summary": "News corrections.publish",
        "operationId": "corrections.publish",
        "parameters": [
          {
            "name": "correctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NewsCorrectionNotice"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-news",
        "x-sdkwork-api-authority": "sdkwork-news.backend",
        "x-sdkwork-domain": "news",
        "x-sdkwork-resource": "corrections",
        "x-sdkwork-public": false
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "NewsItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "categoryId",
          "slug",
          "title",
          "summary",
          "status",
          "priority"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/NewsItemStatus"
          },
          "authorName": {
            "type": "string"
          },
          "featured": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer"
          },
          "estimatedReadMinutes": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledFor": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewsBreakingAlert": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "title",
          "summary",
          "severity",
          "audienceType",
          "priority",
          "status",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "breaking",
              "important",
              "watch"
            ]
          },
          "audienceType": {
            "type": "string",
            "enum": [
              "all",
              "subscribers",
              "targeted"
            ]
          },
          "targetType": {
            "type": "string",
            "enum": [
              "source",
              "author",
              "topic",
              "channel",
              "tag"
            ]
          },
          "targetId": {
            "type": "string"
          },
          "priority": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "cancelled",
              "expired"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewsDigestIssue": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "digestKey",
          "title",
          "digestType",
          "status",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "digestKey": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "digestType": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "topic",
              "editorial"
            ]
          },
          "audienceType": {
            "type": "string",
            "enum": [
              "all",
              "subscribers",
              "targeted"
            ]
          },
          "locale": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "archived"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewsFactCheck": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "claim",
          "verdict",
          "summary",
          "status",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "claim": {
            "type": "string"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "true",
              "mostly_true",
              "mixed",
              "mostly_false",
              "false",
              "unverified"
            ]
          },
          "summary": {
            "type": "string"
          },
          "evidenceUrl": {
            "type": "string",
            "format": "uri"
          },
          "reviewerUserId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewsCorrectionNotice": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "itemId",
          "correctionType",
          "title",
          "body",
          "status",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "correctionType": {
            "type": "string",
            "enum": [
              "correction",
              "clarification",
              "retraction",
              "update"
            ]
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "actorUserId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewsItemStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "scheduled",
          "archived"
        ]
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          }
        }
      },
...
    }
  },
...
}